refactor: move the transaction dialog flow out of the design system
`dialog.tsx` in @sushiswap/ui held two unrelated things: generic Radix
wrappers, and a two-step review -> confirm transaction flow. The second is
app domain logic, not a UI primitive:

- DialogConfirm takes a chainId and a txHash and resolves explorer links
  via getChainById, so the design system carried chain knowledge.
- confirm() hardcodes the Review -> Confirm transition.
- DialogType caps the flow at exactly two steps, named after a swap flow.

The flow now lives in apps/web/src/lib/transaction-dialog (DialogProvider,
DialogType, useDialog, DialogReview, DialogConfirm). @sushiswap/ui keeps
only the Radix wrappers and their cva variants, and dialog.tsx drops its
imports of sushi, the check/failed mark icons, and the '../index' barrel
(it now needs just IconButton and classnames).

The 22 call sites split their imports: primitives still come from
@sushiswap/ui, the flow from src/lib/transaction-dialog. No behaviour
changes; the components moved verbatim.

Two changes beyond the move:
- DialogContentProps is now exported from @sushiswap/ui, since the app's
  DialogConfirm extends it.
- UseDialog<T> becomes an intersection instead of a conditional type, so
  `open` and `setOpen` stay reachable while T is still generic. The
  previous form could not be consumed from a generic component.

apps/storybook/stories/dialogconfirm.stories.tsx is removed: storybook
documents the design system and only depends on @sushiswap/ui, so it can
no longer reach this flow. The generic dialog.stories.tsx stays.

Note this does not remove `sushi` from @sushiswap/ui — 11 other files in
the package still import it.
