+++ apps/web/src/app/(networks)/(evm)/[chainId]/(trade)/cross-chain-swap/_ui/lifi/trade-review-dialog/hooks/use-evm-cross-chain-swap-trade-review.ts
@@ -1,13 +1,13 @@
 'use client'
 
 import { SwapEventName, sendAnalyticsEvent } from '@sushiswap/telemetry'
-import { DialogType, useDialog } from '@sushiswap/ui'
 import { useEffect, useMemo } from 'react'
 import type { LifiXSwapSupportedChainId } from 'src/config'
 import { APPROVE_TAG_XSWAP } from 'src/lib/constants'
 import { useCrossChainTradeStep } from 'src/lib/hooks/react-query/cross-chain-trade/use-cross-chain-trade-step'
 import { logger } from 'src/lib/logger'
 import { getCrossChainFeesBreakdown } from 'src/lib/swap/cross-chain'
+import { DialogType, useDialog } from 'src/lib/transaction-dialog'
 import { isUserRejectedError } from 'src/lib/wagmi/errors'
 import { useApproved } from 'src/lib/wagmi/systems/checker/provider'
 import { waitForSuccessfulReceipt } from 'src/lib/wagmi/transactions/wait-for-successful-receipt'
