+++ apps/web/src/app/(networks)/(evm)/[chainId]/(trade)/swap/_ui/simple-swap-trade-button.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { DialogTrigger, DialogType, Message, useDialog } from '@sushiswap/ui'
+import { DialogTrigger, Message } from '@sushiswap/ui'
 import { Button } from '@sushiswap/ui'
 import type React from 'react'
 import { type FC, useEffect, useMemo, useState } from 'react'
@@ -11,6 +11,7 @@ import { APPROVE_TAG_SWAP } from 'src/lib/constants'
 import { usePersistedSlippageError } from 'src/lib/hooks/use-persisted-slippage-error'
 import { useSlippageTolerance } from 'src/lib/hooks/use-slippage-tolerance'
 import { warningSeverity } from 'src/lib/swap/warning-severity'
+import { DialogType, useDialog } from 'src/lib/transaction-dialog'
 import { Amounts } from 'src/lib/wagmi/systems/checker/amounts'
 import { ApproveERC20 } from 'src/lib/wagmi/systems/checker/approve-erc20'
 import { Connect } from 'src/lib/wagmi/systems/checker/connect'
