+++ apps/web/src/app/(networks)/(evm)/[chainId]/(trade)/swap/_ui/simple-swap-trade-review-dialog/use-svm-simple-swap-trade-review.ts
@@ -5,13 +5,13 @@ import { useTransactionSigner } from '@solana/connector'
 import type { Base64EncodedWireTransaction } from '@solana/kit'
 import { createErrorToast, createToast } from '@sushiswap/notifications'
 import { SwapEventName, sendAnalyticsEvent } from '@sushiswap/telemetry'
-import { DialogType, useDialog } from '@sushiswap/ui'
 import { useCallback, useMemo, useRef, useState } from 'react'
 import type { SupportedChainId } from 'src/config'
 import { sendDrilldownLog } from 'src/lib/drilldown-log'
 import type { UseSvmTradeReturn } from 'src/lib/hooks/react-query'
 import { useSvmTradeExecute } from 'src/lib/hooks/react-query/trade/use-svm-trade-execute'
 import { waitForSvmSignature } from 'src/lib/svm/wait-for-svm-signature'
+import { DialogType, useDialog } from 'src/lib/transaction-dialog'
 import { useAccount } from 'src/lib/wallet/hooks/use-account'
 import { Amount, Percent } from 'sushi'
 import {
